Create README.md
👋 Hi, I’m [Your Name]

I'm an IT graduate passionate about starting my career in cybersecurity. With a solid foundation in IT, I'm diving into the world of cybersecurity to protect digital landscapes and secure information. My journey includes learning and applying skills in penetration testing, threat analysis, and security scripting. I'm enthusiastic about contributing to projects, collaborating with the community, and constantly expanding my knowledge in this dynamic field.